// src/types/PlacemarkService/PlacemarkPointInfo.ts
interface PlacemarkPointInfo {
  default_pixel_size: number;
  default_color: string;
  default_outline_color: string;
  default_outline_width: number;
  highlight_pixel_size: number;
  highlight_color: string;
  highlight_outline_color: string;
  highlight_outline_width: number;
}

export type { PlacemarkPointInfo };